Add new changes again...

See merge request mishakov/pmos!7

+++ kernel/arch/loongarch64/loongarch_asm.hh
@@ -68,29 +68,29 @@ inline void invalidate_user_page(void *addr, unsigned asid)
     asm volatile ("invtlb 0x5, %0, %1" :: "r"(asid), "r"(addr) : "memory");
 }
 
-inline void iocsr_write8(u8 value, u32 address)
+inline void iocsr_write8(u8 value, ulong address)
 {
     asm volatile ("iocsrwr.b %0, %1" :: "r"(value), "r"(address) : "memory");
 }
 
-inline void iocsr_write32(u32 value, u32 address)
+inline void iocsr_write32(u32 value, ulong address)
 {
     asm volatile ("iocsrwr.w %0, %1" :: "r"(value), "r"(address) : "memory");
 }
 
-inline void iocsr_write64(u64 value, u32 address)
+inline void iocsr_write64(u64 value, ulong address)
 {
     asm volatile ("iocsrwr.d %0, %1" :: "r"(value), "r"(address) : "memory");
 }
 
-inline u32 iocsr_read32(u32 address)
+inline u32 iocsr_read32(ulong address)
 {
     u32 value;
     asm volatile ("iocsrrd.w %0, %1" : "=r"(value) : "r"(address));
     return value;
 }
 
-inline u64 iocsr_read64(u32 address)
+inline u64 iocsr_read64(ulong address)
 {
     u64 value;
     asm volatile ("iocsrrd.d %0, %1" : "=r"(value) : "r"(address));
